requirements.txt
Django==1.9.6
facebook-sdk==1.0.0
requests==2.10.0
